intermediate_codes/web_development_projects/spotify_musical_time_machine/requirements.txt
requests==2.32.5
beautifulsoup4==4.14.2
python-dotenv==1.0.0
spotipy==2.23.0